Allow all tests to use bugsnag_server fixture
